requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
decorator==5.0.9
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-decouple==3.4
selenium==3.141.0
soupsieve==2.2.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
